+++ test/unit/mysql_rep_cmp/run_program.py
@@ -28,7 +28,7 @@ import version                                  # pylint:disable=E0401,C0413
 __version__ = version.__version__
 
 
-class ArgParser():                                      # pylint:disable=R0903
+class ArgParser():
 
     """Class:  ArgParser
 
@@ -37,6 +37,7 @@ class ArgParser():                                      # pylint:disable=R0903
     Methods:
         __init__
         get_val
+        arg_exist
 
     """
 
@@ -64,6 +65,18 @@ class ArgParser():                                      # pylint:disable=R0903
         """
 
         return self.args_array.get(skey, def_val)
+
+    def arg_exist(self, arg):
+
+        """Method:  arg_exist
+
+        Description:  Method stub holder for gen_class.ArgParser.arg_exist.
+
+        Arguments:
+
+        """
+
+        return  arg in self.args_array
 
 
 class Mail():                                           # pylint:disable=R0903
@@ -211,6 +224,7 @@ class UnitTest(unittest.TestCase):
 
     Methods:
         setUp
+        test_non_slave_compare
         test_mysql_version3
         test_mysql_version2
         test_mysql_version
@@ -242,33 +256,41 @@ class UnitTest(unittest.TestCase):
         self.master = MasterRep()
         self.slave = SlaveRep()
         self.sys_ign_db = ["performance_schema", "information_schema"]
+
         self.args = ArgParser()
         self.args2 = ArgParser()
-        self.args2a = ArgParser()
-        self.args3 = ArgParser()
-        self.args3a = ArgParser()
-        self.args4 = ArgParser()
         self.args5 = ArgParser()
         self.args6 = ArgParser()
+        self.args7 = ArgParser()
+
         self.args.args_array = {"-c": True, "-d": True, "-r": True}
         self.args2.args_array = {
-            "-c": True, "-d": True, "-r": True, "-e": "email_address",
-            "-s": "subject_line"}
-        self.args2a.args_array = {
-            "-c": True, "-d": True, "-r": True, "-e": "email_address",
-            "-s": "subject_line", "-u": True}
-        self.args3.args_array = {
-            "-c": True, "-d": True, "-r": True, "-e": "email_address"}
-        self.args3a.args_array = {
-            "-c": True, "-d": True, "-r": True, "-e": "email_address",
-            "-u": True}
-        self.args4.args_array = {
-            "-c": True, "-d": True, "-r": True, "-z": True}
+            "-c": True, "-d": True, "-r": True, "-i": True}
         self.args5.args_array = {
             "-c": True, "-d": True, "-r": True, "-t": ["tbl1", "tbl2"],
             "-B": "db1"}
         self.args6.args_array = {
             "-c": True, "-d": True, "-r": True, "-B": "db1"}
+
+    @mock.patch("mysql_rep_cmp.setup_cmp", mock.Mock(return_value=True))
+    @mock.patch("mysql_rep_cmp.mysql_libs.disconnect",
+                mock.Mock(return_value=True))
+    @mock.patch("mysql_rep_cmp.mysql_libs.create_instance")
+    def test_non_slave_compare(self, mock_server):
+
+        """Function:  test_non_slave_compare
+
+        Description:  Test with a non-slave database.
+
+        Arguments:
+
+        """
+
+        self.slave.server_id = 12
+
+        mock_server.side_effect = [self.master, self.slave]
+
+        self.assertFalse(mysql_rep_cmp.run_program(self.args2))
 
     @mock.patch("mysql_rep_cmp.setup_cmp", mock.Mock(return_value=True))
     @mock.patch("mysql_rep_cmp.mysql_libs.disconnect",